# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PéAPInière API",
        "description": "Documentation de l'API pour la pépinière",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/plants": {
            "get": {
                "tags": [
                    "Plantes"
                ],
                "summary": "Lister toutes les plantes",
                "operationId": "6e2a3a91967b1e364c87b755cd5152f6",
                "responses": {
                    "200": {
                        "description": "Liste des plantes avec images et catégorie"
                    }
                }
            },
            "post": {
                "tags": [
                    "Plantes"
                ],
                "summary": "Créer une nouvelle plante",
                "operationId": "d005e00370ab9bd321d50ba5dde92042",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "category_id",
                                    "images"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Basilic aromatique"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Une plante aromatique très utilisée."
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 12.5
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "https://example.com/basilic.jpg"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plante créée avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation (ex. + de 4 images)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/plants/{slug}": {
            "get": {
                "tags": [
                    "Plantes"
                ],
                "summary": "Afficher une plante par son slug",
                "operationId": "cf101271037916b9c4a3e8bed0a3bf4e",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug de la plante",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plante trouvée"
                    },
                    "404": {
                        "description": "Plante non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Plantes"
                ],
                "summary": "Modifier une plante existante",
                "operationId": "e5981af9ce55cca7fc8eee04c3c0b488",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug de la plante à modifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Menthe fraîche"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Parfumée et fraîche."
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 10
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "https://example.com/menthe1.jpg"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plante mise à jour"
                    },
                    "422": {
                        "description": "Nombre d'images dépasse 4"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Plantes"
                ],
                "summary": "Supprimer une plante",
                "operationId": "72abaa41b86e9d3386648d6c79c2b1ab",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug de la plante",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plante supprimée"
                    },
                    "404": {
                        "description": "Plante non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/statistics/total-orders": {
            "get": {
                "tags": [
                    "Statistiques"
                ],
                "summary": "Nombre total de commandes",
                "description": "Utilise la méthode Query Builder :\n     * - `DB::table('orders')` : sélectionne la table `orders`\n     * - `->count()` : compte le nombre total de lignes (commandes)",
                "operationId": "126367309951974d5a6fd24b527c8f27",
                "responses": {
                    "200": {
                        "description": "Nombre total de commandes"
                    },
                    "404": {
                        "description": "Aucune commande trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/statistics/total-revenue": {
            "get": {
                "tags": [
                    "Statistiques"
                ],
                "summary": "Montant total des ventes",
                "description": "Calcule le chiffre d'affaires total :\n     * - `DB::table('order_plant')` : accède à la table pivot\n     * - `->join('plants', ...)` : jointure pour accéder au prix de chaque plante\n     * - `->select(DB::raw('SUM(order_plant.quantity * plants.price) as total'))` : somme des ventes\n     * - `->value('total')` : retourne le total directement",
                "operationId": "303b679ac70802176253feb45155bb38",
                "responses": {
                    "200": {
                        "description": "Montant total des ventes"
                    },
                    "404": {
                        "description": "Aucune vente enregistrée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/statistics/top-plants": {
            "get": {
                "tags": [
                    "Statistiques"
                ],
                "summary": "Top 5 des plantes les plus commandées",
                "description": "Retourne les plantes les plus populaires :\n     * - `DB::table('order_plant')` : table pivot\n     * - `->join('plants', ...)` : accède aux noms des plantes\n     * - `->select('plants.name', DB::raw('SUM(order_plant.quantity) as total_quantity'))` : total commandé par plante\n     * - `->groupBy('plants.name')` : groupement par nom\n     * - `->orderByDesc('total_quantity')` : tri décroissant\n     * - `->limit(5)` : limite aux 5 premières",
                "operationId": "62281b97863015c9434929a9ab89dee8",
                "responses": {
                    "200": {
                        "description": "Liste des 5 plantes les plus populaires"
                    },
                    "404": {
                        "description": "Aucune plante commandée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/statistics/sales-by-category": {
            "get": {
                "tags": [
                    "Statistiques"
                ],
                "summary": "Répartition des ventes par catégorie",
                "description": "Calcule les ventes par catégorie :\n     * - Jointure sur `order_plant`, `plants`, `categories`\n     * - `SUM(order_plant.quantity * plants.price)` : total des ventes par catégorie\n     * - `groupBy('categories.name')` : groupement par catégorie",
                "operationId": "51d57171778d3e139a7d0898072e56b3",
                "responses": {
                    "200": {
                        "description": "Liste des ventes par catégorie"
                    },
                    "404": {
                        "description": "Aucune vente enregistrée par catégorie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT pour accéder aux routes protégées",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}